feat: Refactored the mechanism for sync, so now both server and client will know that each is done syncing. Now conflicts will be handled by accepting the latest one. Client changes will be automatically sent to the client now

+++ pkg/models/item.go
@@ -11,7 +11,7 @@ type Item struct {
 	// ServerPath is the relative to the user vault file path
 	ServerPath string `json:"server_path" form:"server_path" binding:"required" bson:"server_path"`
 	// ServerMTime contains the last time the file has had a change
-	ServerMTime int `json:"server_m_time" form:"server_m_time" binding:"required" bson:"server_m_time"`
+	ServerMTime int64 `json:"server_m_time" form:"server_m_time" binding:"required" bson:"server_m_time"`
 	// SHA256 contains the server caluclated SHA256 of the file
 	SHA256 string `json:"sha256" form:"sha256" binding:"required" bson:"sha256"`
 	// Size contains the bytes size of the file.
